+++ pkg/dao/types/resource_revision.go
@@ -11,3 +11,12 @@ type OutputValue struct {
 	Value  property.Value  `json:"value,omitempty"`
 	Schema openapi3.Schema `json:"schema,omitempty"`
 }
+
+const (
+	ResourceRevisionTypeApply   = "apply"
+	ResourceRevisionTypeDestory = "destroy"
+	// ResourceRevisionTypeSync try to sync the resource from remote and update the revision state to match.
+	ResourceRevisionTypeSync = "sync"
+	// ResourceRevisionTypeDetect try to detect drift of the resource from remote.
+	ResourceRevisionTypeDetect = "detect"
+)
