+++ pkg/deployer/terraform/context.go
@@ -4,8 +4,6 @@ import (
 	"github.com/seal-io/walrus/pkg/dao/types/object"
 )
 
-const WalrusContextVariableName = "context"
-
 // Context indicate the walrus related metadata, will set to attribute context while user module include this attribute.
 type Context struct {
 	// Project indicate the project metadata.
