+++ pkg/deployer/terraform/deployer.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"regexp"
-	"strings"
 
 	"github.com/hashicorp/terraform-config-inspect/tfconfig"
 	corev1 "k8s.io/api/core/v1"
@@ -13,14 +12,11 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	apiconfig "github.com/seal-io/walrus/pkg/apis/config"
-	"github.com/seal-io/walrus/pkg/auths"
 	"github.com/seal-io/walrus/pkg/auths/session"
 	revisionbus "github.com/seal-io/walrus/pkg/bus/resourcerevision"
 	"github.com/seal-io/walrus/pkg/dao"
 	"github.com/seal-io/walrus/pkg/dao/model"
-	"github.com/seal-io/walrus/pkg/dao/model/connector"
 	"github.com/seal-io/walrus/pkg/dao/model/environment"
-	"github.com/seal-io/walrus/pkg/dao/model/environmentconnectorrelationship"
 	"github.com/seal-io/walrus/pkg/dao/model/predicate"
 	"github.com/seal-io/walrus/pkg/dao/model/project"
 	"github.com/seal-io/walrus/pkg/dao/model/resource"
@@ -40,29 +36,10 @@ import (
 	pkgresource "github.com/seal-io/walrus/pkg/resource"
 	"github.com/seal-io/walrus/pkg/resourcedefinitions"
 	"github.com/seal-io/walrus/pkg/settings"
-	"github.com/seal-io/walrus/pkg/templates/openapi"
-	"github.com/seal-io/walrus/pkg/templates/translator"
 	"github.com/seal-io/walrus/pkg/terraform/config"
 	"github.com/seal-io/walrus/pkg/terraform/parser"
 	"github.com/seal-io/walrus/pkg/terraform/util"
-	"github.com/seal-io/walrus/utils/json"
 	"github.com/seal-io/walrus/utils/log"
-	"github.com/seal-io/walrus/utils/pointer"
-)
-
-// DeployerType the type of deployer.
-const DeployerType = types.DeployerTypeTF
-
-const (
-	// _backendAPI the API path to terraform deploy backend.
-	// Terraform will get and update deployment states from this API.
-	_backendAPI = "/v1/projects/%s/environments/%s/resources/%s/revisions/%s/terraform-states"
-
-	// _variablePrefix the prefix of the variable name.
-	_variablePrefix = "_walrus_var_"
-
-	// _resourcePrefix the prefix of the service output name.
-	_resourcePrefix = "_walrus_res_"
 )
 
 var (
@@ -84,6 +61,31 @@ type Deployer struct {
 	logger      log.Logger
 	modelClient model.ClientSet
 	clientSet   *kubernetes.Clientset
+}
+
+type createRevisionOptions struct {
+	// ResourceID indicates the ID of resource which is for create the revision.
+	ResourceID object.ID
+	// JobType indicates the type of the job.
+	JobType string
+}
+
+type createK8sJobOptions struct {
+	// Type indicates the type of the job.
+	Type string
+	// ResourceRevision indicates the resource revision to create the deployment job.
+	ResourceRevision *model.ResourceRevision
+}
+
+type createK8sSecretsOptions struct {
+	ResourceRevision *model.ResourceRevision
+	Connectors       model.Connectors
+	SubjectID        object.ID
+	// Walrus Context.
+	Context Context
+
+	Token     string
+	ServerURL string
 }
 
 func NewDeployer(_ context.Context, opts deptypes.CreateOptions) (deptypes.Deployer, error) {
@@ -126,10 +128,12 @@ func (d Deployer) Apply(ctx context.Context, resource *model.Resource, opts dept
 		_ = d.updateRevisionStatus(ctx, revision)
 	}()
 
-	return d.createK8sJob(ctx, createK8sJobOptions{
+	err = d.createK8sJob(ctx, createK8sJobOptions{
 		Type:             JobTypeApply,
 		ResourceRevision: revision,
 	})
+
+	return err
 }
 
 // Destroy creates a new resource revision by the given resource,
@@ -155,24 +159,84 @@ func (d Deployer) Destroy(ctx context.Context, resource *model.Resource, opts de
 		_ = d.updateRevisionStatus(ctx, revision)
 	}()
 
-	return d.createK8sJob(ctx, createK8sJobOptions{
+	err = d.createK8sJob(ctx, createK8sJobOptions{
 		Type:             JobTypeDestroy,
 		ResourceRevision: revision,
 	})
+
+	return err
 }
 
-type createK8sJobOptions struct {
-	// Type indicates the type of the job.
-	Type string
-	// ResourceRevision indicates the resource revision to create the deployment job.
-	ResourceRevision *model.ResourceRevision
+// Sync creates a new resource revision by the given resource,
+// and drives the Kubernetes Job to sync the components of the resource.
+func (d Deployer) Sync(ctx context.Context, resource *model.Resource, opts deptypes.SyncOptions) (err error) {
+	revision, err := d.createRevision(ctx, createRevisionOptions{
+		ResourceID: resource.ID,
+		JobType:    JobTypeSync,
+	})
+	if err != nil {
+		return err
+	}
+
+	defer func() {
+		if err == nil {
+			return
+		}
+
+		// Update a failure status.
+		status.ResourceRevisionStatusReady.False(revision, err.Error())
+
+		// Report to resource revision.
+		_ = d.updateRevisionStatus(ctx, revision)
+	}()
+
+	err = d.createK8sJob(ctx, createK8sJobOptions{
+		Type:             JobTypeSync,
+		ResourceRevision: revision,
+	})
+
+	return err
+}
+
+// Detect will detect resource changes from remote system of given service.
+func (d Deployer) Detect(
+	ctx context.Context,
+	resource *model.Resource,
+	opts deptypes.DetectOptions,
+) error {
+	revision, err := d.createRevision(ctx, createRevisionOptions{
+		ResourceID: resource.ID,
+		JobType:    JobTypeDetect,
+	})
+	if err != nil {
+		return err
+	}
+
+	defer func() {
+		if err == nil {
+			return
+		}
+
+		// Update a failure status.
+		status.ResourceRevisionStatusReady.False(revision, err.Error())
+
+		// Report to resource revision.
+		_ = d.updateRevisionStatus(ctx, revision)
+	}()
+
+	err = d.createK8sJob(ctx, createK8sJobOptions{
+		Type:             JobTypeDetect,
+		ResourceRevision: revision,
+	})
+
+	return err
 }
 
 // createK8sJob creates a k8s job to deploy, destroy or rollback the resource.
 func (d Deployer) createK8sJob(ctx context.Context, opts createK8sJobOptions) error {
 	revision := opts.ResourceRevision
 
-	connectors, err := d.getConnectors(ctx, revision.EnvironmentID)
+	connectors, err := dao.GetEnvironmentConnectors(ctx, d.modelClient, revision.EnvironmentID)
 	if err != nil {
 		return err
 	}
@@ -219,6 +283,17 @@ func (d Deployer) createK8sJob(ctx context.Context, opts createK8sJobOptions) er
 			SetEnvironment(env.ID, env.Name, pkgenv.GetManagedNamespaceName(env)).
 			SetResource(res.ID, res.Name),
 	}
+
+	secretOpts.ServerURL, err = getServerURL(ctx, d.modelClient)
+	if err != nil {
+		return err
+	}
+
+	secretOpts.Token, err = getToken(ctx, d.modelClient, secretOpts.SubjectID, opts.ResourceRevision.ID)
+	if err != nil {
+		return err
+	}
+
 	if err = d.createK8sSecrets(ctx, secretOpts); err != nil {
 		return err
 	}
@@ -235,13 +310,14 @@ func (d Deployer) createK8sJob(ctx context.Context, opts createK8sJobOptions) er
 
 	// Create deployment job.
 	jobOpts := JobCreateOptions{
-		Type:               opts.Type,
-		ResourceRevisionID: opts.ResourceRevision.ID.String(),
-		Image:              jobImage,
-		Env:                jobEnv,
+		Type:      opts.Type,
+		Image:     jobImage,
+		Env:       jobEnv,
+		ServerURL: secretOpts.ServerURL,
+		Token:     secretOpts.Token,
 	}
 
-	return CreateJob(ctx, d.clientSet, jobOpts)
+	return createJob(ctx, d.clientSet, opts.ResourceRevision, jobOpts)
 }
 
 func (d Deployer) getEnv(ctx context.Context) ([]corev1.EnvVar, error) {
@@ -324,14 +400,6 @@ func (d Deployer) updateRevisionStatus(ctx context.Context, ar *model.ResourceRe
 	return nil
 }
 
-type createK8sSecretsOptions struct {
-	ResourceRevision *model.ResourceRevision
-	Connectors       model.Connectors
-	SubjectID        object.ID
-	// Walrus Context.
-	Context Context
-}
-
 // createK8sSecrets creates the k8s secrets for deployment.
 func (d Deployer) createK8sSecrets(ctx context.Context, opts createK8sSecretsOptions) error {
 	secretData := make(map[string][]byte)
@@ -364,13 +432,6 @@ func (d Deployer) createK8sSecrets(ctx context.Context, opts createK8sSecretsOpt
 	}
 
 	return nil
-}
-
-type createRevisionOptions struct {
-	// ResourceID indicates the ID of resource which is for create the revision.
-	ResourceID object.ID
-	// JobType indicates the type of the job.
-	JobType string
 }
 
 // createRevision creates a new resource revision.
@@ -497,6 +558,7 @@ func (d Deployer) createRevision(
 	}
 
 	entity := &model.ResourceRevision{
+		Type:            opts.JobType,
 		ProjectID:       res.ProjectID,
 		EnvironmentID:   res.EnvironmentID,
 		ResourceID:      res.ID,
@@ -586,7 +648,6 @@ func (d Deployer) getRequiredProviders(
 
 // loadConfigsBytes returns terraform main.tf and terraform.tfvars for deployment.
 func (d Deployer) loadConfigsBytes(ctx context.Context, opts createK8sSecretsOptions) (map[string][]byte, error) {
-	logger := log.WithName("deployer").WithName("tf")
 	// Prepare terraform tfConfig.
 	//  get module configs from resource revision.
 	moduleConfig, providerRequirements, err := d.getModuleConfig(ctx, opts)
@@ -602,7 +663,7 @@ func (d Deployer) loadConfigsBytes(ctx context.Context, opts createK8sSecretsOpt
 		if _, ok := requiredProviders[p.Name]; !ok {
 			requiredProviders[p.Name] = p.ProviderRequirement
 		} else {
-			logger.Warnf("duplicate provider requirement: %s", p.Name)
+			d.logger.Warnf("duplicate provider requirement: %s", p.Name)
 		}
 	}
 
@@ -632,30 +693,15 @@ func (d Deployer) loadConfigsBytes(ctx context.Context, opts createK8sSecretsOpt
 		return nil, err
 	}
 
-	// Prepare address for terraform backend.
-	serverAddress, err := settings.ServeUrl.Value(ctx, d.modelClient)
-	if err != nil {
-		return nil, err
-	}
-
-	if serverAddress == "" {
+	if opts.ServerURL == "" {
 		return nil, errors.New("server address is empty")
 	}
-	address := fmt.Sprintf("%s%s", serverAddress,
+	address := fmt.Sprintf("%s%s", opts.ServerURL,
 		fmt.Sprintf(_backendAPI,
 			opts.Context.Project.ID,
 			opts.Context.Environment.ID,
 			opts.Context.Resource.ID,
 			opts.ResourceRevision.ID))
-
-	// Prepare API token for terraform backend.
-	const _1Day = 60 * 60 * 24
-
-	at, err := auths.CreateAccessToken(ctx,
-		d.modelClient, opts.SubjectID, types.TokenKindDeployment, string(opts.ResourceRevision.ID), pointer.Int(_1Day))
-	if err != nil {
-		return nil, err
-	}
 
 	// Prepare terraform config files to be mounted to secret.
 	requiredProviderNames := sets.NewString()
@@ -666,7 +712,7 @@ func (d Deployer) loadConfigsBytes(ctx context.Context, opts createK8sSecretsOpt
 	secretOptionMaps := map[string]config.CreateOptions{
 		config.FileMain: {
 			TerraformOptions: &config.TerraformOptions{
-				Token:                at.AccessToken,
+				Token:                opts.Token,
 				Address:              address,
 				SkipTLSVerify:        !apiconfig.TlsCertified.Get(),
 				ProviderRequirements: requiredProviders,
@@ -793,31 +839,6 @@ func (d Deployer) getModuleConfig(
 	return mc, requiredProviders, err
 }
 
-func (d Deployer) getConnectors(ctx context.Context, environmentID object.ID) (model.Connectors, error) {
-	rs, err := d.modelClient.EnvironmentConnectorRelationships().Query().
-		Where(environmentconnectorrelationship.EnvironmentID(environmentID)).
-		WithConnector(func(cq *model.ConnectorQuery) {
-			cq.Select(
-				connector.FieldID,
-				connector.FieldName,
-				connector.FieldType,
-				connector.FieldCategory,
-				connector.FieldConfigVersion,
-				connector.FieldConfigData)
-		}).
-		All(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	var cs model.Connectors
-	for i := range rs {
-		cs = append(cs, rs[i].Edges.Connector)
-	}
-
-	return cs, nil
-}
-
 // getPreviousRequiredProviders get previous succeed revision required providers.
 // NB(alex): the previous revision may be failed, the failed revision may not contain required providers of states.
 func (d Deployer) getPreviousRequiredProviders(
@@ -853,180 +874,4 @@ func (d Deployer) getPreviousRequiredProviders(
 	}
 
 	return prevRequiredProviders, nil
-}
-
-func getVarConfigOptions(
-	variables model.Variables,
-	resourceOutputs map[string]parser.OutputState,
-) config.CreateOptions {
-	varsConfigOpts := config.CreateOptions{
-		Attributes: map[string]any{},
-	}
-
-	for _, v := range variables {
-		varsConfigOpts.Attributes[_variablePrefix+v.Name] = v.Value
-	}
-
-	// Setup resource outputs.
-	for n, v := range resourceOutputs {
-		varsConfigOpts.Attributes[_resourcePrefix+n] = v.Value
-	}
-
-	return varsConfigOpts
-}
-
-func getModuleConfig(
-	revision *model.ResourceRevision,
-	template *model.TemplateVersion,
-	opts createK8sSecretsOptions,
-) (*config.ModuleConfig, error) {
-	mc := &config.ModuleConfig{
-		Name:   opts.Context.Resource.Name,
-		Source: template.Source,
-	}
-
-	if template.Schema.IsEmpty() {
-		return mc, nil
-	}
-
-	mc.SchemaData = template.Schema.TemplateVersionSchemaData
-
-	if template.Schema.OpenAPISchema == nil ||
-		template.Schema.OpenAPISchema.Components == nil ||
-		template.Schema.OpenAPISchema.Components.Schemas == nil {
-		return mc, nil
-	}
-
-	// Variables.
-	var (
-		variablesSchema    = template.Schema.VariableSchema()
-		outputsSchemas     = template.Schema.OutputSchema()
-		sensitiveVariables = sets.Set[string]{}
-	)
-
-	if variablesSchema != nil {
-		attrs, err := translator.ToGoTypeValues(revision.Attributes, *variablesSchema)
-		if err != nil {
-			return nil, err
-		}
-
-		mc.Attributes = attrs
-
-		for n, v := range variablesSchema.Properties {
-			// Add sensitive from schema variable.
-			if v.Value.WriteOnly {
-				sensitiveVariables.Insert(fmt.Sprintf(`var\.%s`, n))
-			}
-
-			if n == WalrusContextVariableName {
-				mc.Attributes[n] = opts.Context
-			}
-		}
-	}
-
-	// Outputs.
-	if outputsSchemas != nil {
-		sps := outputsSchemas.Properties
-		mc.Outputs = make([]config.Output, 0, len(sps))
-
-		sensitiveVariableRegex, err := matchAnyRegex(sensitiveVariables.UnsortedList())
-		if err != nil {
-			return nil, err
-		}
-
-		for k, v := range sps {
-			origin := openapi.GetExtOriginal(v.Value.Extensions)
-			co := config.Output{
-				Sensitive:    v.Value.WriteOnly,
-				Name:         k,
-				ResourceName: opts.Context.Resource.Name,
-				Value:        origin.ValueExpression,
-			}
-
-			if !v.Value.WriteOnly {
-				// Update sensitive while output is from sensitive data, like secret.
-				if sensitiveVariables.Len() != 0 && sensitiveVariableRegex.Match(origin.ValueExpression) {
-					co.Sensitive = true
-				}
-			}
-
-			mc.Outputs = append(mc.Outputs, co)
-		}
-	}
-
-	return mc, nil
-}
-
-func updateOutputWithVariables(variables model.Variables, moduleConfig *config.ModuleConfig) (map[string]bool, error) {
-	var (
-		variableOpts         = make(map[string]bool)
-		encryptVariableNames = sets.NewString()
-	)
-
-	for _, s := range variables {
-		variableOpts[s.Name] = s.Sensitive
-
-		if s.Sensitive {
-			encryptVariableNames.Insert(_variablePrefix + s.Name)
-		}
-	}
-
-	if encryptVariableNames.Len() == 0 {
-		return variableOpts, nil
-	}
-
-	reg, err := matchAnyRegex(encryptVariableNames.UnsortedList())
-	if err != nil {
-		return nil, err
-	}
-
-	var shouldEncryptAttr []string
-
-	for k, v := range moduleConfig.Attributes {
-		b, err := json.Marshal(v)
-		if err != nil {
-			return nil, err
-		}
-
-		matches := reg.FindAllString(string(b), -1)
-		if len(matches) != 0 {
-			shouldEncryptAttr = append(shouldEncryptAttr, fmt.Sprintf(`var\.%s`, k))
-		}
-	}
-
-	// Outputs use encrypted variable should set to sensitive.
-	for i, v := range moduleConfig.Outputs {
-		if v.Sensitive {
-			continue
-		}
-
-		reg, err := matchAnyRegex(shouldEncryptAttr)
-		if err != nil {
-			return nil, err
-		}
-
-		if reg.MatchString(string(v.Value)) {
-			moduleConfig.Outputs[i].Sensitive = true
-		}
-	}
-
-	return variableOpts, nil
-}
-
-func matchAnyRegex(list []string) (*regexp.Regexp, error) {
-	var sb strings.Builder
-
-	sb.WriteString("(")
-
-	for i, v := range list {
-		sb.WriteString(v)
-
-		if i < len(list)-1 {
-			sb.WriteString("|")
-		}
-	}
-
-	sb.WriteString(")")
-
-	return regexp.Compile(sb.String())
 }
