+++ pkg/deployer/types/types.go
@@ -21,6 +21,12 @@ type Deployer interface {
 
 	// Destroy cleans all resources of the given resource.
 	Destroy(context.Context, *model.Resource, DestroyOptions) error
+
+	// Sync syncs the state of the given resource.
+	Sync(context.Context, *model.Resource, SyncOptions) error
+
+	// Detect detect drift of the given resource.
+	Detect(context.Context, *model.Resource, DetectOptions) error
 }
 
 // ApplyOptions holds the options of Deployer's Apply action.
@@ -28,3 +34,9 @@ type ApplyOptions struct{}
 
 // DestroyOptions holds the options of Deployer's Destroy action.
 type DestroyOptions struct{}
+
+// SyncOptions holds the options of Deployer's Sync action.
+type SyncOptions struct{}
+
+// DeployerFactory creates a Deployer.
+type DetectOptions struct{}
