+++ pkg/resource/resource.go
@@ -4,16 +4,10 @@ import (
 	"context"
 	"fmt"
 
-	"entgo.io/ent/dialect/sql"
-	"entgo.io/ent/dialect/sql/sqljson"
-
 	"github.com/seal-io/walrus/pkg/auths/session"
 	"github.com/seal-io/walrus/pkg/dao"
 	"github.com/seal-io/walrus/pkg/dao/model"
-	"github.com/seal-io/walrus/pkg/dao/model/resource"
 	"github.com/seal-io/walrus/pkg/dao/model/resourcecomponent"
-	"github.com/seal-io/walrus/pkg/dao/model/resourcerelationship"
-	"github.com/seal-io/walrus/pkg/dao/types"
 	"github.com/seal-io/walrus/pkg/dao/types/object"
 	"github.com/seal-io/walrus/pkg/dao/types/status"
 	deptypes "github.com/seal-io/walrus/pkg/deployer/types"
@@ -133,19 +127,10 @@ func Destroy(
 		return mc.Resources().DeleteOneID(entity.ID).Exec(ctx)
 	}
 
-	updateFailedStatus := func(err error) {
-		status.ResourceStatusDeleted.False(entity, err.Error())
-
-		err = UpdateStatus(ctx, mc, entity)
-		if err != nil {
-			logger.Errorf("error updating status of resource %s: %v", entity.ID, err)
-		}
-	}
-
 	// Check dependants.
 	dependants, err := dao.GetResourceDependantNames(ctx, mc, entity)
 	if err != nil {
-		updateFailedStatus(err)
+		updateFailedStatus(ctx, mc, entity, status.ResourceStatusDeleted, err)
 		return err
 	}
 
@@ -175,9 +160,9 @@ func Destroy(
 
 	err = opts.Deployer.Destroy(ctx, entity, deptypes.DestroyOptions{})
 	if err != nil {
-		log.Errorf("fail to destroy resource: %w", err)
+		logger.Errorf("fail to destroy resource: %w", err)
 
-		updateFailedStatus(err)
+		updateFailedStatus(ctx, mc, entity, status.ResourceStatusDeleted, err)
 	}
 
 	return nil
@@ -192,19 +177,10 @@ func Stop(
 ) (err error) {
 	logger := log.WithName("resource")
 
-	updateFailedStatus := func(err error) {
-		status.ResourceStatusStopped.False(entity, err.Error())
-
-		err = UpdateStatus(ctx, mc, entity)
-		if err != nil {
-			logger.Errorf("error updating status of resource %s: %v", entity.ID, err)
-		}
-	}
-
 	// Check dependants.
 	dependants, err := dao.GetResourceDependantNames(ctx, mc, entity, status.ResourceStatusStopped.String())
 	if err != nil {
-		updateFailedStatus(err)
+		updateFailedStatus(ctx, mc, entity, status.ResourceStatusStopped, err)
 		return err
 	}
 
@@ -234,9 +210,59 @@ func Stop(
 
 	err = opts.Deployer.Destroy(ctx, entity, deptypes.DestroyOptions{})
 	if err != nil {
-		log.Errorf("fail to destroy resource: %w", err)
+		logger.Errorf("fail to destroy resource: %w", err)
 
-		updateFailedStatus(err)
+		updateFailedStatus(ctx, mc, entity, status.ResourceStatusStopped, err)
+	}
+
+	return nil
+}
+
+func Sync(
+	ctx context.Context,
+	mc model.ClientSet,
+	entity *model.Resource,
+	opts Options,
+) (err error) {
+	logger := log.WithName("resource")
+
+	// Mark status to syncing.
+	status.ResourceStatusSynced.Reset(entity, "")
+
+	if err = UpdateStatus(ctx, mc, entity); err != nil {
+		return fmt.Errorf("failed to update resource status: %w", err)
+	}
+
+	err = opts.Deployer.Sync(ctx, entity, deptypes.SyncOptions{})
+	if err != nil {
+		logger.Errorf("fail to sync resource: %w", err)
+
+		updateFailedStatus(ctx, mc, entity, status.ResourceStatusSynced, err)
+	}
+
+	return nil
+}
+
+func Detect(
+	ctx context.Context,
+	mc model.ClientSet,
+	entity *model.Resource,
+	opts Options,
+) (err error) {
+	logger := log.WithName("resource")
+
+	// Mark status to detecting.
+	status.ResourceStatusDetected.Reset(entity, "")
+
+	if err = UpdateStatus(ctx, mc, entity); err != nil {
+		return fmt.Errorf("failed to update resource status: %w", err)
+	}
+
+	err = opts.Deployer.Detect(ctx, entity, deptypes.DetectOptions{})
+	if err != nil {
+		logger.Errorf("fail to detect resource: %w", err)
+
+		updateFailedStatus(ctx, mc, entity, status.ResourceStatusDetected, err)
 	}
 
 	return nil
@@ -374,132 +400,19 @@ func CreateScheduledResources(
 	return results, nil
 }
 
-// IsStatusReady returns true if the resource is ready.
-func IsStatusReady(entity *model.Resource) bool {
-	switch entity.Status.SummaryStatus {
-	case "Preparing", "NotReady", "Ready":
-		return true
-	}
+func updateFailedStatus(
+	ctx context.Context,
+	mc model.ClientSet,
+	entity *model.Resource,
+	status status.ConditionType,
+	failErr error,
+) {
+	logger := log.WithName("resource")
 
-	return false
-}
+	status.False(entity, failErr.Error())
 
-// IsStatusFalse returns true if the resource is in error status.
-func IsStatusFalse(entity *model.Resource) bool {
-	switch entity.Status.SummaryStatus {
-	case "DeployFailed", "DeleteFailed":
-		return true
-	case "Progressing":
-		return entity.Status.Error
-	}
-
-	return false
-}
-
-// IsStatusDeleted returns true if the resource is deleted.
-func IsStatusDeleted(entity *model.Resource) bool {
-	switch entity.Status.SummaryStatus {
-	case "Deleted", "Deleting":
-		return true
-	}
-
-	return false
-}
-
-const (
-	summaryStatusDeploying   = "Deploying"
-	summaryStatusProgressing = "Progressing"
-)
-
-// CheckDependencyStatus check resource dependencies status is ready to apply.
-func CheckDependencyStatus(ctx context.Context, mc model.ClientSet, entity *model.Resource) (bool, error) {
-	// Check dependants.
-	dependencies, err := mc.ResourceRelationships().Query().
-		Where(
-			resourcerelationship.ResourceID(entity.ID),
-			resourcerelationship.DependencyIDNEQ(entity.ID),
-		).
-		QueryDependency().
-		Select(resource.FieldID).
-		Where(
-			resource.Or(
-				func(s *sql.Selector) {
-					s.Where(sqljson.ValueEQ(
-						resource.FieldStatus,
-						summaryStatusDeploying,
-						sqljson.Path("summaryStatus"),
-					))
-				},
-				resource.And(
-					func(s *sql.Selector) {
-						s.Where(sqljson.ValueEQ(
-							resource.FieldStatus,
-							summaryStatusProgressing,
-							sqljson.Path("summaryStatus"),
-						))
-					},
-					func(s *sql.Selector) {
-						s.Where(sqljson.ValueEQ(
-							resource.FieldStatus,
-							true,
-							sqljson.Path("transitioning"),
-						))
-					},
-				),
-			),
-		).
-		All(ctx)
+	err := UpdateStatus(ctx, mc, entity)
 	if err != nil {
-		return false, err
+		logger.Errorf("error updating status of resource %s: %v", entity.ID, err)
 	}
-
-	if len(dependencies) > 0 {
-		// If dependency resources is in deploying status.
-		err = SetResourceStatusScheduled(ctx, mc, entity)
-		if err != nil {
-			return false, err
-		}
-
-		return false, nil
-	}
-
-	return true, nil
-}
-
-// IsService tells if the given resource is of service type.
-func IsService(r *model.Resource) bool {
-	if r == nil {
-		return false
-	}
-
-	return r.TemplateID != nil
-}
-
-// IsStoppable tells whether the given resource is stoppable.
-func IsStoppable(r *model.Resource) bool {
-	if r == nil {
-		return false
-	}
-
-	if r.Labels[types.LabelResourceStoppable] == "true" ||
-		(r.TemplateID != nil && r.Labels[types.LabelResourceStoppable] != "false") {
-		return true
-	}
-
-	return false
-}
-
-// CanBeStopped tells whether the given resource can be stopped.
-func CanBeStopped(r *model.Resource) bool {
-	return status.ResourceStatusDeployed.IsTrue(r)
-}
-
-// IsInactive tells whether the given resource is inactive.
-func IsInactive(r *model.Resource) bool {
-	if r == nil {
-		return false
-	}
-
-	return r.Status.SummaryStatus == status.ResourceStatusUnDeployed.String() ||
-		r.Status.SummaryStatus == status.ResourceStatusStopped.String()
 }
